Added projects tiles to project page with basic styling
