+++ src/App.tsx
@@ -6,7 +6,7 @@ import AboutPage from './pages/about/about'
 import NotFoundPage from './pages/not_found'
 import ContactPage from './pages/contact'
 import StackPage from './pages/stack/stack'
-import ProjectsPage from '../src/pages/projects'
+import ProjectsPage from './pages/projects/projects'
 import PrivacyPage from '../src/pages/privacy'
 import CookiesPage from '../src/pages/cookies'
 import PlcCarrierPage from './pages/about/plc_carrier'
