+++ src/language/pl.json
@@ -333,6 +333,9 @@
         },
         "stack":{
             "front_header": "Umiejętności Front-end developera"
+        },
+        "projects":{
+            "header": "Projekty"
         }
     }
 }
