--- src/pages/projects.tsx
@@ -1,16 +1,0 @@
-import TypedText from '../components/typedText'
-import {setPage} from '../reducers/state'
-import {useDispatch } from 'react-redux'
-import Breadcrumbs from '../components/breadcrumbs'
-
-const ProjectsPage = () => {
-  const dispatch = useDispatch();
-  dispatch(setPage("Projects"));
-  return(
-    <div style={{ display: 'flex', flexDirection: 'column', alignItems: 'center', height: '100%'}}>
-      <Breadcrumbs />
-      <TypedText />
-    </div>
-  );
-};
-export default ProjectsPage;
