+++ src/styles/pages/projects.scss
@@ -1,0 +1,37 @@
+.projects-wrapper {
+    width: 80vw;
+    display: grid;
+    grid-template-columns: calc(40vw - 0px) calc(40vw - 0px);
+    grid-template-rows: calc(40vw - 0px) calc(40vw - 0px);
+    justify-items: center;
+    align-items: center;
+    //column-gap: 20px;
+    //row-gap: 20px;
+    border: 1px solid #fff;
+    margin: 15px;
+    box-sizing: border-box;
+    @media (max-width: 968px){
+        width: 90vw;
+        grid-template-columns: calc(45vw - 0px) calc(45vw - 0px);
+        grid-template-rows: calc(45vw - 0px) calc(45vw - 0px);
+    }
+    @media (max-width: 668px){
+        width: 80vw;
+        grid-template-columns: 100%;
+        grid-template-rows: auto auto auto auto;
+        row-gap: 20px;
+    }
+  }
+  .projects-hello{
+    font-size: 82px;
+    text-align: center;
+    @media (max-width: 768px){
+        font-size: 50px;
+    }
+    margin: 15px auto;
+    font-family: 'Times New Roman', serif;
+}
+.projects-hello::after{
+    content: ".";
+    color: rgb(30, 229, 255);
+}
